Print values in dictionary
A dictionary in Python is an unordered set of key: value pairs. Unlike lists, dictionaries are inherently orderless. The key: value pairs appear to be in a certain order but it is irrelevant. Each KEY must be unique, but the VALUES may be the same for two or more keys.